Add files via upload
 # Sparse Matrix using linked lists

The aim of this work is to implement and test a particular form of representation to store and manipulate sparse matrices: linked lists. They are opposed to other formats that are also usual to represent sparse matrices, typically based on sets of index and value vectors (such as CRS, CCS, etc.). These formats provide an efficient way to implement usual mathematical operations on arrays, but they are not as efficient for representing structures with variable and / or unknown sizes.
It is in this last type of scenario that the representation by linked lists presents advantages.